added Dev Activity
I added an activity called DevActivity. This can be used by us to help
debug and test things, while developing the app. For example: I am
working on developing a specific screen and I will add a button to the
Dev Page that would direct me the screen I want.
On the Home screen, press Menu and select "Dev Page".