Background and menu image are all set
If you guys come across anything that you think would improve the look
of our app, let me know and I'll try to get it in there. Just shoot me a
link. Do we want to change the SplashScreen? At the most I think we
would just change the colors so it is consistent, but other than that, I
think it looks fine.

Doug